+++ js/card.js
@@ -1,4 +1,25 @@
-import { offerData } from './data.js';
+const offerData = {
+  'palace': {
+    label: 'Дворец',
+    price: 10000,
+  },
+  'flat': {
+    label: 'Квартира',
+    price: 1000,
+  },
+  'bungalow': {
+    label: 'Бунгало',
+    price: 0,
+  },
+  'house': {
+    label: 'Дом',
+    price: 5000,
+  },
+  'hotel': {
+    label: 'Отель',
+    price: 3000,
+  },
+};
 
 const cardTemplate = document
   .querySelector('#card')
@@ -86,5 +107,6 @@ const createCard = (data) => {
 };
 
 export {
-  createCard
+  createCard,
+  offerData
 };
