--- js/data.js
@@ -1,26 +1,0 @@
-const offerData = {
-  'palace': {
-    label: 'Дворец',
-    price: 10000,
-  },
-  'flat': {
-    label: 'Квартира',
-    price: 1000,
-  },
-  'bungalow': {
-    label: 'Бунгало',
-    price: 0,
-  },
-  'house': {
-    label: 'Дом',
-    price: 5000,
-  },
-  'hotel': {
-    label: 'Отель',
-    price: 3000,
-  },
-};
-
-export {
-  offerData
-};
